chore: general cleanup and use 128 bits perps account

+++ src/markets/perps/interfaces/IPerpsAccountModule.sol
@@ -17,7 +17,7 @@ interface IPerpsAccountModule {
     /// @notice Emitted when a new perps account is created.
     /// @param accountId The trading account id.
     /// @param sender The `msg.sender` of the create account transaction.
-    event LogCreatePerpsAccount(uint256 accountId, address sender);
+    event LogCreatePerpsAccount(uint128 accountId, address sender);
 
     /// @notice Emitted when `msg.sender` deposits `amount` of `collateralType` into `accountId`.
     /// @param sender The `msg.sender`.
@@ -46,7 +46,7 @@ interface IPerpsAccountModule {
     /// @param collateralType The margin collateral address.
     /// @return marginCollateralBalance The margin amount of the given collateral type.
     function getAccountMarginCollateralBalance(
-        uint256 accountId,
+        uint128 accountId,
         address collateralType
     )
         external
@@ -57,7 +57,7 @@ interface IPerpsAccountModule {
     /// @dev This function doesn't take open positions into account.
     /// @param accountId The trading account id.
     /// @return totalMarginCollateralValue The USD denominated total margin collateral value.
-    function getTotalAccountMarginCollateralValue(uint256 accountId)
+    function getTotalAccountMarginCollateralValue(uint128 accountId)
         external
         view
         returns (UD60x18 totalMarginCollateralValue);
@@ -71,14 +71,14 @@ interface IPerpsAccountModule {
     /// @return availableMargin The account's withdrawable margin balance.
     /// @return initialMargin The account's initial margin in positions.
     /// @return maintenanceMargin The account's maintenance margin.
-    function getAccountMarginBalances(uint256 accountId)
+    function getAccountMarginBalances(uint128 accountId)
         external
         view
         returns (SD59x18 marginBalance, SD59x18 availableMargin, UD60x18 initialMargin, UD60x18 maintenanceMargin);
 
     /// @notice Creates a new trading account and mints its NFT
     /// @return accountId The trading account id.
-    function createPerpsAccount() external returns (uint256 accountId);
+    function createPerpsAccount() external returns (uint128 accountId);
 
     /// @notice Creates a new trading account and multicalls using the provided data payload.
     /// @param data The data payload to be multicalled.
@@ -89,13 +89,13 @@ interface IPerpsAccountModule {
     /// @param accountId The trading account id.
     /// @param collateralType The margin collateral address.
     /// @param amount The amount of margin collateral to deposit.
-    function depositMargin(uint256 accountId, address collateralType, uint256 amount) external;
+    function depositMargin(uint128 accountId, address collateralType, uint256 amount) external;
 
     /// @notice Withdraws available margin collateral from the given trading account.
     /// @param accountId The trading account id.
     /// @param collateralType The margin collateral address.
     /// @param amount The amount of margin collateral to withdraw.
-    function withdrawMargin(uint256 accountId, address collateralType, uint256 amount) external;
+    function withdrawMargin(uint128 accountId, address collateralType, uint256 amount) external;
 
     /// @notice Used by the Account NFT contract to notify an account transfer.
     /// @dev Can only be called by the Account NFT contract.
