+++ src/markets/perps/modules/PerpsMarketModule.sol
@@ -83,7 +83,7 @@ abstract contract PerpsMarketModule is IPerpsMarketModule {
         fillPrice = perpsMarket.getIndexPrice();
     }
 
-    function getPositionLeverage(uint256 accountId, uint128 marketId) external view override returns (UD60x18) {
+    function getPositionLeverage(uint128 accountId, uint128 marketId) external view override returns (UD60x18) {
         PerpsMarket.Data storage perpsMarket = PerpsMarket.load(marketId);
         Position.Data storage position = perpsMarket.positions[accountId];
 
@@ -122,7 +122,7 @@ abstract contract PerpsMarketModule is IPerpsMarketModule {
 
     /// @inheritdoc IPerpsMarketModule
     function getOpenPositionData(
-        uint256 accountId,
+        uint128 accountId,
         uint128 marketId
     )
         external
